use refactored column component
