+++ app/javascript/mastodon/features/follow_requests/index.jsx
@@ -11,14 +11,15 @@ import { connect } from 'react-redux';
 import { debounce } from 'lodash';
 
 import PersonAddIcon from '@/material-icons/400-24px/person_add.svg?react';
+import { Column } from '@/mastodon/components/column';
 import { injectIntl } from '@/mastodon/components/intl';
 
 import { fetchFollowRequests, expandFollowRequests } from '../../actions/accounts';
 import ScrollableList from '../../components/scrollable_list';
 import { me } from '../../initial_state';
-import Column from '../ui/components/column';
 
 import AccountAuthorizeContainer from './containers/account_authorize_container';
+import { ColumnHeader } from '@/mastodon/components/column/header';
 
 const messages = defineMessages({
   heading: { id: 'column.follow_requests', defaultMessage: 'Follow requests' },
@@ -69,7 +70,8 @@ class FollowRequests extends ImmutablePureComponent {
     );
 
     return (
-      <Column bindToDocument={!multiColumn} icon='user-plus' iconComponent={PersonAddIcon} heading={intl.formatMessage(messages.heading)} alwaysShowBackButton>
+      <Column bindToDocument={!multiColumn}>
+        <ColumnHeader icon='user-plus' iconComponent={PersonAddIcon} title={intl.formatMessage(messages.heading)} showBackButton />
         <ScrollableList
           scrollKey='follow_requests'
           onLoadMore={this.handleLoadMore}
