+++ app/javascript/mastodon/features/mutes/index.jsx
@@ -11,13 +11,14 @@ import { connect } from 'react-redux';
 import { debounce } from 'lodash';
 
 import VolumeOffIcon from '@/material-icons/400-24px/volume_off.svg?react';
-import { Account } from 'mastodon/components/account';
+import { fetchMutes, expandMutes } from '@/mastodon/actions/mutes';
+import { Account } from '@/mastodon/components/account';
+import { Column } from '@/mastodon/components/column';
 import { injectIntl } from '@/mastodon/components/intl';
+import { LoadingIndicator } from '@/mastodon/components/loading_indicator';
+import ScrollableList from '@/mastodon/components/scrollable_list';
+import { ColumnHeader } from '@/mastodon/components/column/header';
 
-import { fetchMutes, expandMutes } from '../../actions/mutes';
-import { LoadingIndicator } from '../../components/loading_indicator';
-import ScrollableList from '../../components/scrollable_list';
-import Column from '../ui/components/column';
 
 const messages = defineMessages({
   heading: { id: 'column.mutes', defaultMessage: 'Muted users' },
@@ -63,7 +64,8 @@ class Mutes extends ImmutablePureComponent {
     const emptyMessage = <FormattedMessage id='empty_column.mutes' defaultMessage="You haven't muted any users yet." />;
 
     return (
-      <Column bindToDocument={!multiColumn} icon='volume-off' iconComponent={VolumeOffIcon} heading={intl.formatMessage(messages.heading)} alwaysShowBackButton>
+      <Column bindToDocument={!multiColumn}>
+        <ColumnHeader icon='volume-off' iconComponent={VolumeOffIcon} title={intl.formatMessage(messages.heading)} showBackButton />
         <ScrollableList
           scrollKey='mutes'
           onLoadMore={this.handleLoadMore}
