+++ app/javascript/mastodon/features/reblogs/index.jsx
@@ -11,15 +11,14 @@ import { connect } from 'react-redux';
 import { debounce } from 'lodash';
 
 import RefreshIcon from '@/material-icons/400-24px/refresh.svg?react';
-import { Account } from 'mastodon/components/account';
+import { fetchReblogs, expandReblogs } from '@/mastodon/actions/interactions';
+import { Account } from '@/mastodon/components/account';
+import { Column } from '@/mastodon/components/column';
 import { ColumnHeader } from '@/mastodon/components/column/header';
-import { Icon }  from 'mastodon/components/icon';
+import { Icon }  from '@/mastodon/components/icon';
 import { injectIntl } from '@/mastodon/components/intl';
-
-import { fetchReblogs, expandReblogs } from '../../actions/interactions';
-import { LoadingIndicator } from '../../components/loading_indicator';
-import ScrollableList from '../../components/scrollable_list';
-import Column from '../ui/components/column';
+import { LoadingIndicator } from '@/mastodon/components/loading_indicator';
+import ScrollableList from '@/mastodon/components/scrollable_list';
 
 const messages = defineMessages({
   refresh: { id: 'refresh', defaultMessage: 'Refresh' },
